Move navigation to left to make space for future expansion
