+++ ui/src/components/layout/Navigation.js
@@ -1,0 +1,44 @@
+import React, { PropTypes } from 'react';
+
+import { Link } from 'react-router';
+import { Menu, Container } from 'semantic-ui-react';
+
+export default class TopNav extends React.Component {
+  static propTypes = {
+    signOut: PropTypes.func.isRequired
+  };
+
+  render() {
+    const { signOut, username, location } = this.props;
+    return (
+      <Menu fixed="left" inverted vertical color="blue">
+        <Container>
+          <Menu.Item header>
+            Shipyard
+          </Menu.Item>
+
+          <Menu.Item>
+            <div className="header">Swarm</div>
+            <Menu.Menu>
+              <Menu.Item name="services" icon="cubes" to="/services" as={Link} active={location.pathname === "/services" || location.pathname === "/"} />
+              <Menu.Item name="containers" icon="cube" to="/containers" as={Link} active={location.pathname === "/containers"} />
+              <Menu.Item name="images" icon="file" to="/images" as={Link} active={location.pathname === "/images"} />
+              <Menu.Item name="nodes" icon="server" to="/nodes" as={Link} active={location.pathname === "/nodes"} />
+              <Menu.Item name="networks" icon="sitemap" to="/networks" as={Link} active={location.pathname === "/networks"} />
+              <Menu.Item name="volumes" icon="database" to="/volumes" as={Link} active={location.pathname === "/volumes"} />
+              <Menu.Item name="accounts" icon="users" to="/accounts" as={Link} active={location.pathname === "/accounts"} />
+              <Menu.Item name="settings" icon="settings" to="/settings" as={Link} active={location.pathname === "/settings"} />
+            </Menu.Menu>
+          </Menu.Item>
+
+          <Menu.Item>
+            <div className="header">{username}</div>
+            <Menu.Menu>
+              <Menu.Item icon="sign out" name="sign out" onClick={signOut}></Menu.Item>
+            </Menu.Menu>
+          </Menu.Item>
+        </Container>
+      </Menu>
+    )
+  }
+}
